Refactor code to facilitate caching
