+++ src/main/java/com/flagsmith/FlagsmithClient.java
@@ -20,13 +20,8 @@ import java.util.List;
  */
 public class FlagsmithClient {
 
-    private FlagsmithConfig defaultConfig;
-    private static final String AUTH_HEADER = "X-Environment-Key";
-    private static final String ACCEPT_HEADER = "Accept";
-    // an api key per environment
-    private String apiKey;
     private final FlagsmithLogger logger = new FlagsmithLogger();
-    private HashMap<String, String> customHeaders;
+    private FlagsmithEndpoints flagsmithEndpoints;
 
     private FlagsmithClient() {
     }
@@ -58,34 +53,7 @@ public class FlagsmithClient {
      * @return a list of feature flags
      */
     public List<Flag> getFeatureFlags(FeatureUser user, boolean doThrow) {
-        HttpUrl.Builder urlBuilder;
-        if (user == null) {
-            urlBuilder = defaultConfig.flagsURI.newBuilder()
-                    .addEncodedQueryParameter("page", "1");
-        } else {
-            urlBuilder = defaultConfig.flagsURI.newBuilder("")
-                    .addEncodedPathSegment(user.getIdentifier());
-        }
-
-        final Request request = this.newRequestBuilder()
-                .url(urlBuilder.build())
-                .build();
-
-        Call call = defaultConfig.httpClient.newCall(request);
-        List<Flag> featureFlags = new ArrayList<>();
-        try (Response response = call.execute()) {
-            if (response.isSuccessful()) {
-                ObjectMapper mapper = MapperFactory.getMappper();
-                featureFlags = Arrays.asList(mapper.readValue(response.body().string(),
-                        Flag[].class));
-            } else {
-                logger.httpError(request, response, doThrow);
-            }
-        } catch (IOException io) {
-            logger.httpError(request, io, doThrow);
-        }
-        logger.info("Got feature flags for user = {}, flags = {}", user, featureFlags);
-        return featureFlags;
+        return this.flagsmithEndpoints.getFeatureFlags(user, doThrow);
     }
 
     /**
@@ -315,29 +283,7 @@ public class FlagsmithClient {
      * @return a list of user Traits and Flags
      */
     public FlagsAndTraits getUserFlagsAndTraits(FeatureUser user, boolean doThrow) {
-        HttpUrl url = defaultConfig.identitiesURI.newBuilder("")
-                .addEncodedQueryParameter("identifier", user.getIdentifier())
-                .build();
-
-        final Request request = this.newRequestBuilder()
-                .url(url)
-                .build();
-
-        Call call = defaultConfig.httpClient.newCall(request);
-
-        FlagsAndTraits flagsAndTraits = new FlagsAndTraits();
-        try (Response response = call.execute()) {
-            if (response.isSuccessful()) {
-                ObjectMapper mapper = MapperFactory.getMappper();
-                flagsAndTraits = mapper.readValue(response.body().string(), FlagsAndTraits.class);
-            } else {
-                logger.httpError(request, response, doThrow);
-            }
-        } catch (IOException io) {
-            logger.httpError(request, io, doThrow);
-        }
-        logger.info("Got feature flags & traits for user = {}, flagsAndTraits = {}", user, flagsAndTraits);
-        return flagsAndTraits;
+        return this.flagsmithEndpoints.getUserFlagsAndTraits(user, doThrow);
     }
 
     /**
@@ -360,7 +306,7 @@ public class FlagsmithClient {
      * @return a Trait object or null if does not exist
      */
     public Trait updateTrait(FeatureUser user, Trait toUpdate, boolean doThrow) {
-        return postUserTraits(user, toUpdate, doThrow);
+        return this.flagsmithEndpoints.postUserTraits(user, toUpdate, doThrow);
     }
 
     /**
@@ -393,94 +339,18 @@ public class FlagsmithClient {
      * @return a list of added Trait objects
      */
     public List<Trait> identifyUserWithTraits(FeatureUser user, List<Trait> traits, boolean doThrow) {
-        // we are using identities endpoint to create bulk user Trait
-        HttpUrl url = defaultConfig.identitiesURI;
-
-        if (user == null || (user.getIdentifier() == null || user.getIdentifier().length() < 1)) {
-            throw new IllegalArgumentException("Missing user Identifier");
-        }
-
-        IdentityTraits identityTraits = new IdentityTraits();
-        identityTraits.setIdentifier(user.getIdentifier());
-        if (traits != null) {
-            identityTraits.setTraits(traits);
-        }
-
-        MediaType JSON = MediaType.parse("application/json; charset=utf-8");
-        RequestBody body = RequestBody.create(JSON, identityTraits.toString());
-
-        final Request request = this.newRequestBuilder()
-                .post(body)
-                .url(url)
-                .build();
-
-        List<Trait> traitsData = new ArrayList<>();
-        Call call = defaultConfig.httpClient.newCall(request);
-        try (Response response = call.execute()) {
-            if (response.isSuccessful()) {
-                ObjectMapper mapper = MapperFactory.getMappper();
-                FlagsAndTraits flagsAndTraits = mapper.readValue(response.body().string(), FlagsAndTraits.class);
-
-                traitsData = flagsAndTraits.getTraits();
-            } else {
-                logger.httpError(request, response, doThrow);
-            }
-        } catch (IOException io) {
-            logger.httpError(request, io, doThrow);
-        }
-        logger.info("Got traits for user = {}, traits = {}", user, traitsData);
-        return traitsData;
+        return flagsmithEndpoints.identifyUserWithTraits(user, traits, doThrow);
     }
-
-    private Trait postUserTraits(FeatureUser user, Trait toUpdate, boolean doThrow) {
-        HttpUrl url = defaultConfig.traitsURI;
-        toUpdate.setIdentity(user);
-
-        MediaType JSON = MediaType.parse("application/json; charset=utf-8");
-        RequestBody body = RequestBody.create(JSON, toUpdate.toString());
-
-        Request request = this.newRequestBuilder()
-                .post(body)
-                .url(url)
-                .build();
-
-        Trait trait = null;
-        Call call = defaultConfig.httpClient.newCall(request);
-        try (Response response = call.execute()) {
-            if (response.isSuccessful()) {
-                ObjectMapper mapper = MapperFactory.getMappper();
-                trait = mapper.readValue(response.body().string(), Trait.class);
-            } else {
-                logger.httpError(request, response, doThrow);
-            }
-        } catch (IOException io) {
-            logger.httpError(request, io, doThrow);
-        }
-        logger.info("Updated trait for user = {}, new trait = {}, updated trait = {}", user, toUpdate, trait);
-        return trait;
-    }
-
 
     public static FlagsmithClient.Builder newBuilder() {
         return new FlagsmithClient.Builder();
     }
 
-    private Request.Builder newRequestBuilder() {
-        final Request.Builder builder = new Request.Builder()
-            .header(AUTH_HEADER, apiKey)
-            .addHeader(ACCEPT_HEADER, "application/json");
-
-        if (this.customHeaders != null && !this.customHeaders.isEmpty()) {
-            this.customHeaders.forEach((k, v) -> builder.addHeader(k, v));
-        }
-
-        return builder;
-    }
-
-
     public static class Builder {
         private FlagsmithClient client;
         private FlagsmithConfig configuration = FlagsmithConfig.newBuilder().build();
+        private HashMap<String, String> customHeaders;
+        private String apiKey;
 
         private Builder() {
             client = new FlagsmithClient();
@@ -496,7 +366,7 @@ public class FlagsmithClient {
             if (null == apiKey) {
                 throw new IllegalArgumentException("Api key can not be null");
             } else {
-                client.apiKey = apiKey;
+                this.apiKey = apiKey;
                 return this;
             }
         }
@@ -557,13 +427,13 @@ public class FlagsmithClient {
          * @return the Builder
          */
         public Builder withCustomHttpHeaders(HashMap<String, String> customHeaders) {
-            this.client.customHeaders = customHeaders;
+            this.customHeaders = customHeaders;
             return this;
         }
 
         public FlagsmithClient build() {
-            client.defaultConfig = this.configuration;
-            return client;
+            this.client.flagsmithEndpoints = new FlagsmithEndpoints(this.configuration, this.customHeaders, client.logger, apiKey);
+            return this.client;
         }
     }
 }
